+++ src/Result.Net/ResultObjects/Result.WithValue.cs
@@ -24,7 +24,13 @@
         /// </summary>
         /// <param name="data">the data associated with the result</param>
         /// <param name="status">the status of the result</param>
-        public Result(TData data, ResultStatus status) : base(status) => Data = data;
+        /// <param name="message">the message associated with the result</param>
+        /// <param name="code">the code associated with the result</param>
+        /// <param name="logTraceCode">the log trace code associated with the result</param>
+        /// <param name="errors">the list of errors if any</param>
+        /// <param name="metaData">the collection of meta-data if any</param>
+        public Result(TData data, ResultStatus status, string message, string code, string logTraceCode, ICollection<ResultError> errors, IDictionary<string, object> metaData)
+            : base(status, message, code, logTraceCode, errors, metaData) => Data = data;
 
         /// <summary>
         /// check if the operation associated with this result has produce a data.