+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0]
+
+### Added
+
+- added a `Cast<>()` extension method that allow you to cast a `Result` instance into a `Result{TData}` instance.
+
+### Changed
+
+- Removed `ListResult<>` no need for this type, to simply the usage the `Result<data[]>` can be used instead.
+- Removed `Clone()` methods, the new `Cast<>()` method should be used instead.
+- All properties on the `Result` instance will not be set and they will have their default value (null, or empty string).
+
+### Fixed
+
+- Fixed JSON Deserialization issue by removing all constructors from the `Result` class and only set one default constructor 
+
+## [1.4.0]
+
+### Added
+
+- added localization support, by introducing a new extension method `WithLocalizedMessage()` on the Result type to allow setting localized messages on the Result object instance.
+
 ## [1.3.0]
 
 ### Added